+++ README.md
@@ -69,7 +69,7 @@ http://localhost:8080/swagger-ui/
 
 - [x] Rest url construction needs some work as it seems to be contradicting the controllers (in which those urls are defined) name, has verbs...
 
-- [x] InterestApplicationTest seems to be testing nothing.
+- [x] InterestApplicationSpec seems to be testing nothing.
 
 - [x] Unmanaged database changes, they are generated by hibernate. Please consider using Liquibase or Flyway.
