+++ build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'org.springframework.boot' version '2.5.2'
+    id 'org.springframework.boot' version '2.6.4'
     id 'io.spring.dependency-management' version '1.0.11.RELEASE'
     id 'java'
     id 'groovy'
@@ -8,54 +8,62 @@ plugins {
 
 group = 'com.4finance'
 version = '1.0-SNAPSHOT'
+//sourceCompatibility = '17'
 
 repositories {
     mavenCentral()
 }
 
 ext {
-    springBootVersion = '2.5.2'
-    groovy = '3.0.8'
-    lombok = '1.18.20'
-    springfox = '3.0.0'
-    mapstruct = '1.4.2.Final'
-    spock = '2.0-groovy-3.0'
-    lombok_mapstruct_bindings = '0.2.0'
-    testcontainers = '1.16.0'
+    springBootVersion = '2.6.4'
+    groovyVersion = '3.0.10'
+    lombokVersion = '1.18.22'
+    springfoxVersion = '3.0.0'
+    mapstructVersion = '1.4.2.Final'
+    spockVersion = '2.1-groovy-3.0'
+    lombokMapstructBindingsVersion = '0.2.0'
+    testContainersVersion = '1.16.3'
 }
 
 dependencies {
-    implementation group: 'org.projectlombok', name: 'lombok', version: lombok
-    implementation group: 'org.codehaus.groovy', name: 'groovy-all', version: groovy
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
-    implementation 'redis.clients:jedis'
-    implementation("org.liquibase:liquibase-core")
-    implementation group: 'io.springfox', name: 'springfox-boot-starter', version: springfox
-    implementation group: 'org.mapstruct', name: 'mapstruct', version: mapstruct
+    implementation(
+            "org.projectlombok:lombok:${lombokVersion}",
+            "org.codehaus.groovy:groovy-all:${groovyVersion}",
+            'org.springframework.boot:spring-boot-starter-data-jpa',
+            'org.springframework.boot:spring-boot-starter-web',
+            'org.springframework.boot:spring-boot-starter-validation',
+            'org.springframework.boot:spring-boot-starter-data-redis',
+            'redis.clients:jedis',
+            'org.liquibase:liquibase-core',
+            "io.springfox:springfox-boot-starter:${springfoxVersion}",
+            "org.mapstruct:mapstruct:${mapstructVersion}",
+            'org.postgresql:postgresql'
+    )
 
-    compileOnly group: 'org.projectlombok', name: 'lombok', version: lombok
-    compileOnly group: 'org.projectlombok', name: 'lombok-mapstruct-binding', version: lombok_mapstruct_bindings
+    compileOnly(
+            "org.projectlombok:lombok:${lombokVersion}",
+            "org.projectlombok:lombok-mapstruct-binding:${lombokMapstructBindingsVersion}"
 
-    annotationProcessor group: 'org.projectlombok', name: 'lombok', version: lombok
-    annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
-    annotationProcessor group: 'org.mapstruct', name: 'mapstruct-processor', version: mapstruct
+    )
 
-    runtimeOnly 'org.postgresql:postgresql'
+    annotationProcessor(
+            "org.projectlombok:lombok:${lombokVersion}",
+            'org.springframework.boot:spring-boot-configuration-processor',
+            "org.mapstruct:mapstruct-processor:${mapstructVersion}"
+    )
 
-    testImplementation group: 'org.spockframework', name: 'spock-core', version: spock
-    testImplementation group: 'org.spockframework', name: 'spock-spring', version: spock
-    testImplementation "org.springframework.boot:spring-boot-starter-test"
-    testImplementation group: 'org.testcontainers', name: 'testcontainers', version: testcontainers
-    testImplementation group: 'org.testcontainers', name: 'spock', version: testcontainers
-    testImplementation group: 'org.testcontainers', name: 'postgresql', version: testcontainers
-    testImplementation group: 'com.jupiter-tools', name: 'spring-test-redis', version: '0.1'
+    testImplementation(
+            "org.spockframework:spock-core:${spockVersion}",
+            "org.spockframework:spock-spring:${spockVersion}",
+            'org.springframework.boot:spring-boot-starter-test',
+            "org.testcontainers:testcontainers:${testContainersVersion}",
+            "org.testcontainers:spock:${testContainersVersion}",
+            "org.testcontainers:postgresql:${testContainersVersion}",
+            "com.jupiter-tools:spring-test-redis:0.1"
+    )
 
-    testCompileOnly group: 'org.projectlombok', name: 'lombok', version: lombok
-
-    testAnnotationProcessor group: 'org.projectlombok', name: 'lombok', version: lombok
+//    testCompileOnly group: 'org.projectlombok', name: 'lombok', version: lombok
+//    testAnnotationProcessor group: 'org.projectlombok', name: 'lombok', version: lombok
 }
 
 test {
