+++ docker-compose.yml
@@ -1,0 +1,17 @@
+version: "3.9"
+services:
+  redis:
+    image: "redis:alpine"
+  postgres:
+    image: postgres:14.1-alpine
+    restart: always
+    environment:
+      - POSTGRES_USER=root
+      - POSTGRES_PASSWORD=root
+    ports:
+      - '5432:5432'
+    volumes:
+      - db:/var/lib/postgresql/data
+volumes:
+  db:
+    driver: local
